+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.2.1] - 2024-07-18
+
+### Added
+- New 'export' command to export domains to a CSV file
+- 'import' command to import domains from a CSV file
+- Unit tests for both import and export functionalities
+- Updated README.md with information about new commands
+
+### Changed
+- Refactored main.py to include new import and export commands
+- Updated CLI help text to include new commands
+
 ## [0.2.0] - 2024-07-13
 
 ### Added